Fixed a messed up merge and made percentage nullable in Vat (so we can tell when updating when its no discount and when its an empty field (we dont want toupdate percentage))

+++ backend/Server/Models/DTOs/VAT/VatUpdateDTO.cs
@@ -7,5 +7,5 @@ public class VatUpdateDTO
     public string? Name { get; set; }
 
     [Range(0, 100, ErrorMessage = "Percentage must be between 0 and 100")]
-    public decimal? Percentage { get; set; }
+    public float? Percentage { get; set; }
 }
